% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{add_attributes_for_endpoint}
\alias{add_attributes_for_endpoint}
\title{Add ATTRIBUTES_FOR endpoint to API router}
\usage{
add_attributes_for_endpoint(pr)
}
\arguments{
\item{pr}{A plumber router object}
}
\value{
The modified plumber router
}
\description{
Adds a POST endpoint at \verb{/ATTRIBUTES_FOR} that wraps \code{codeminer::ATTRIBUTES_FOR()}.
}
\keyword{internal}
