% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{add_children_endpoint}
\alias{add_children_endpoint}
\title{Add CHILDREN endpoint to API router}
\usage{
add_children_endpoint(pr)
}
\arguments{
\item{pr}{A plumber router object}
}
\value{
The modified plumber router
}
\description{
Adds a POST endpoint at \verb{/CHILDREN} that wraps \code{codeminer::CHILDREN()}.
}
\keyword{internal}
