% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{add_codes_like_endpoint}
\alias{add_codes_like_endpoint}
\title{Add CODES_LIKE endpoint to API router}
\usage{
add_codes_like_endpoint(pr)
}
\arguments{
\item{pr}{A plumber router object}
}
\value{
The modified plumber router
}
\description{
Adds a POST endpoint at \verb{/CODES_LIKE} that wraps \code{codeminer::CODES_LIKE()}.
}
\keyword{internal}
