% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{add_has_attributes_endpoint}
\alias{add_has_attributes_endpoint}
\title{Add HAS_ATTRIBUTES endpoint to API router}
\usage{
add_has_attributes_endpoint(pr)
}
\arguments{
\item{pr}{A plumber router object}
}
\value{
The modified plumber router
}
\description{
Adds a POST endpoint at \verb{/HAS_ATTRIBUTES} that wraps \code{codeminer::HAS_ATTRIBUTES()}.
}
\keyword{internal}
