% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{add_map_endpoint}
\alias{add_map_endpoint}
\title{Add MAP endpoint to API router}
\usage{
add_map_endpoint(pr)
}
\arguments{
\item{pr}{A plumber router object}
}
\value{
The modified plumber router
}
\description{
Adds a POST endpoint at \verb{/MAP} that wraps \code{codeminer::MAP()}.
}
\keyword{internal}
