% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{add_parents_endpoint}
\alias{add_parents_endpoint}
\title{Add PARENTS endpoint to API router}
\usage{
add_parents_endpoint(pr)
}
\arguments{
\item{pr}{A plumber router object}
}
\value{
The modified plumber router
}
\description{
Adds a POST endpoint at \verb{/PARENTS} that wraps \code{codeminer::PARENTS()}.
}
\keyword{internal}
