% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_codeminer_api.R
\name{check_package_available}
\alias{check_package_available}
\title{Check if a package is available}
\usage{
check_package_available(package)
}
\arguments{
\item{package}{Package name to check}
}
\value{
Logical indicating if package is available
}
\description{
Check if a package is available
}
\keyword{internal}
