% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_helpers.R
\name{format_backend_error}
\alias{format_backend_error}
\title{Format backend error as API response}
\usage{
format_backend_error(e, res)
}
\arguments{
\item{e}{Error object from tryCatch}

\item{res}{Plumber response object}
}
\value{
List to be serialized as JSON error response
}
\description{
Format backend error as API response
}
\keyword{internal}
