% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_codeminer_api.R
\name{run_codeminer_api}
\alias{run_codeminer_api}
\title{Run the CodeMiner API}
\usage{
run_codeminer_api(
  host = "127.0.0.1",
  port = 8000,
  background = FALSE,
  docs = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{host}{Host address (default "127.0.0.1")}

\item{port}{Port number (default 8000)}

\item{background}{Logical. Run API in the background? (default FALSE)}

\item{docs}{Logical. Should swagger docs be enabled? (default TRUE)}

\item{quiet}{Logical. Suppress startup messages? (default FALSE)}

\item{...}{Additional arguments passed to plumber::pr_run()}
}
\description{
Run the CodeMiner API
}
